% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Run a quick EDA analysis}
\usage{
eda(df, target)
}
\arguments{
\item{df}{dataframe}

\item{target}{character}
}
\value{
(list) List that contains statistical summaries and a pairplot.
}
\description{
Run a quick EDA analysis
}
\examples{
df <- data.frame(num1 = c(8.5, 8, 9.2, 9.1, 9.4),
                 num2 = c(0.88, 0.93, 0.95 , 0.92 , 0.91),
                 num3 = c(0.46, 0.78, 0.66, 0.69, 0.52),
                 num4 = c(0.082, 0.078, 0.082, 0.085, 0.066),
                 cat1 = c("Good","Okay","Excellent","Terrible","Good"),
                 target = c(2,2,3,1,3))
result <- eda(df,"target")
nb_num_feat  <- result$nb_num_features
nb_cat_feat <- result$nb_cat_features
ls_cat_feat <- result$cat_features
ls_num_feat <- result$num_features
nb_class<- result$nb_classes
pair_plot <- result$pairplot
}
