% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Imputing Missing Data}
\usage{
imputation(fit_data, fill_data, method, constant = NULL)
}
\arguments{
\item{fit_data}{list}

\item{fill_data}{list}

\item{method}{character}

\item{constant}{character, default NULL}
}
\value{
list
}
\description{
This function will impute missing data in a tibble/dataframe given the chosen method(mean, median)
}
\examples{
test_df <- data.frame('a' = c(1,NA,3), 'b' = c(5,6,NA), 'c' = c(NA,1,10))
test_df_imputed <- imputation(test_df, test_df, 'mean')
}
