% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Run a quick EDA analysis}
\usage{
eda(df, target)
}
\arguments{
\item{df}{dataframe}

\item{target}{character}
}
\value{
(list) List that contains statistical summaries and a pairplot.
}
\description{
Run a quick EDA analysis
}
\examples{
result <- eda(mtcars)
nb_num_feat  <- result$nb_num_features
nb_cat_feat <- result$nb_cat_features
ls_cat_feat <- result$cat_features
ls_num_feat <- result$num_features
nb_class<- result$nb_classes
pair_plot <- result$pairplot
}
