% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler}
\alias{scaler}
\title{Scaling a dataframe}
\usage{
scaler(X_train, X_valid, X_test, scaler_type)
}
\arguments{
\item{X_train}{data.frame}

\item{X_valid}{data.frame}

\item{X_test}{data.frame}

\item{scaler_type}{character}
}
\value{
data.frame of data.frames
}
\description{
This function scales numerical features based on scaling requirement(standardization, minmax Scaling) in a data.frame
}
\examples{
scaled_df <- scaler(X_train, X_Valid, X_test, scaler_type='standardization')
}
